requirements.txt
fastapi==0.95.2
starlette==0.27.0
sqlite-utils==3.35.1
langchain==0.0.220
python-dotenv==0.21.0
uvicorn==0.18.3
openai
googletrans==4.0.0-rc1
httpx>=0.13.3<=0.19.0